pyproject.toml
[project]
name = "3dgs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "imageio[ffmpeg]>=2.37.0",
    "jaxtyping==0.2.36",
    "numpy>=2.3.4",
    "pillow>=12.0.0",
    "plyfile>=1.1.3",
    "torchmetrics[image]>=1.8.2",
    "tqdm>=4.67.1",
    "typeguard==2.13.3",
    "tyro>=0.8.0",
]
